pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "openenv-redline_env"
version = "0.1.0"
description = "Redline Env environment for OpenEnv"
requires-python = ">=3.10"
dependencies = [
    "openenv-core[core]>=0.2.2",
    "numpy",
    "pandas",
    "pydantic",
    "fastapi",
    "uvicorn"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
server = "redline_env.server.app:main"

[tool.setuptools]
include-package-data = true
packages = ["redline_env", "redline_env.server"]
package-dir = { "redline_env" = ".", "redline_env.server" = "server" }
